+++ cli/delete.go
@@ -21,6 +21,7 @@ func deleteCommand(t *core.Track) *cobra.Command {
 	}
 
 	edit.AddCommand(deleteRecordCommand(t))
+	edit.AddCommand(deleteProjectCommand(t))
 
 	edit.Long += "\n\n" + formatCmdTree(edit)
 	return edit
@@ -36,12 +37,6 @@ func deleteRecordCommand(t *core.Track) *cobra.Command {
 		Aliases: []string{"r"},
 		Args:    util.WrappedArgs(cobra.ExactArgs(2)),
 		Run: func(cmd *cobra.Command, args []string) {
-			open, ok := t.OpenRecord()
-			if !ok {
-				out.Err("failed to stop record: no record running")
-				return
-			}
-
 			timeString := strings.Join(args, " ")
 			tm, err := util.ParseDateTime(timeString)
 			if err != nil {
@@ -54,7 +49,7 @@ func deleteRecordCommand(t *core.Track) *cobra.Command {
 				return
 			}
 
-			if !force && !confirmDeleteRecord(open) {
+			if !force && !confirmDeleteRecord(record) {
 				out.Err("failed to delete record: aborted by user")
 				return
 			}
@@ -65,6 +60,59 @@ func deleteRecordCommand(t *core.Track) *cobra.Command {
 				return
 			}
 			out.Success("Deleted record %s from '%s'", record.Start.Format(util.DateTimeFormat), record.Project)
+		},
+	}
+
+	delete.Flags().BoolVarP(&force, "force", "F", false, "Don't prompt for confirmation.")
+
+	return delete
+}
+
+func deleteProjectCommand(t *core.Track) *cobra.Command {
+	var force bool
+
+	delete := &cobra.Command{
+		Use:     "project PROJECT",
+		Short:   "Delete a project and all associated records",
+		Long:    "Delete a project and all associated records",
+		Aliases: []string{"p"},
+		Args:    util.WrappedArgs(cobra.ExactArgs(1)),
+		Run: func(cmd *cobra.Command, args []string) {
+			name := args[0]
+
+			projects, err := t.LoadAllProjects()
+			if err != nil {
+				out.Err("failed to delete project: %s", err)
+				return
+			}
+			pTree, err := t.ToProjectTree(projects)
+			if err != nil {
+				out.Err("failed to delete project: %s", err)
+				return
+			}
+
+			pNode, ok := pTree.Nodes[name]
+			if !ok {
+				out.Err("failed to delete project: no project named '%s'", name)
+				return
+			}
+			if len(pNode.Children) > 0 {
+				out.Err("failed to delete project: '%s' has %d child project(s)", name, len(pNode.Children))
+				return
+			}
+
+			if !force && !confirmDeleteProject(pNode.Value) {
+				out.Err("failed to delete project: aborted by user")
+				return
+			}
+
+			cnt, err := t.DeleteProject(pNode.Value)
+			if err != nil {
+				out.Err("failed to delete project: %s", err)
+				out.Err("deleted %d records", cnt)
+				return
+			}
+			out.Success("Deleted project '%s' (%d records)", pNode.Value.Name, cnt)
 		},
 	}