+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Projects can be archived (and un-archived) to exclude them from reports etc. (#39)
 * Project colors and symbols are shown in reports and lists (#40)
 * Weekly sum of project time shown in week report (#41)
+* Delete a project with `track delete project PROJECT` (#43)
 
 ### Bug fixes